do NOT, I repeat do not run the getRandomExcuse or getAllExcuses.
Now a loooong message but here I go;
DataLoader by default is set to add 1.6 Million data...
NOT - we do NOT have pagination and indexing during this commit.
The reason for that is to see the difference before and after indexing and pagination.
With this situation now you will feel deep inside your bones and soul why they are best practice, and you need them. I'm afraid it will crash my Mac. This is some solid reason lmao